@@ -19,7 +19,7 @@
     "@types/react-native": "^0.73.0",
     "babel-plugin-module-resolver": "^4.1.0",
     "burnt": "^0.12.2",
-    "expo": "^51.0.9",
+    "expo": "^53.0.13",
     "expo-constants": "~16.0.2",
     "expo-dev-client": "~4.0.15",
     "expo-font": "~12.0.6",
@@ -30,9 +30,9 @@
     "expo-splash-screen": "0.27.4",
     "expo-status-bar": "~1.12.1",
     "expo-updates": "~0.25.15",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-native": "^0.74.2",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "react-native": "^0.77.2",
     "react-native-gesture-handler": "^2.16.2",
     "react-native-safe-area-context": "^4.10.4",
     "react-native-screens": "^3.31.1",
@@ -42,10 +42,10 @@
   "devDependencies": {
     "@babel/core": "^7.26.9",
     "@expo/metro-config": "^0.18.4",
-    "@tamagui/babel-plugin": "^1.125.14",
+    "@tamagui/babel-plugin": "^1.129.11",
     "babel-plugin-transform-inline-environment-variables": "^0.4.4",
     "metro-minify-terser": "^0.80.9",
-    "typescript": "^5.7.3"
+    "typescript": "^5.8.3"
   },
   "resolutions": {
     "metro": "0.76.0",
