@@ -21,26 +21,26 @@
     "@aws-sdk/credential-provider-node": "^3.750.0",
     "@aws-sdk/lib-dynamodb": "^3.750.0",
     "@my/app": "0.0.1",
-    "@tamagui/config": "^1.125.14",
-    "@tamagui/next-theme": "^1.125.14",
+    "@tamagui/config": "^1.129.11",
+    "@tamagui/next-theme": "^1.129.11",
     "@vercel/analytics": "^1.5.0",
     "bcryptjs": "^3.0.2",
     "google-auth-library": "^9.15.0",
     "google-spreadsheet": "^4.1.4",
-    "next": "14.2.21",
-    "next-auth": "^5.0.0-beta.25",
+    "next": "15.3.4",
+    "next-auth": "^5.0.0-beta.29",
     "nodemailer": "^6.10.0",
     "raf": "^3.4.1",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-native": "^0.74.2",
-    "react-native-web": "^0.19.12",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "react-native": "^0.77.2",
+    "react-native-web": "^0.20.0",
     "react-native-web-lite": "^1.112.4",
     "vercel": "^39.2.2"
   },
   "devDependencies": {
     "@playwright/test": "^1.53.1",
-    "@tamagui/next-plugin": "^1.125.14",
+    "@tamagui/next-plugin": "^1.129.11",
     "@types/bcryptjs": "^3.0.0",
     "@types/node": "^20.14.15",
     "@types/nodemailer": "^6.4.15",
