@@ -6,7 +6,7 @@
     "packages/*"
   ],
   "engines": {
-    "node": "18.18.0",
+    "node": "22.17.0",
     "npm": "please-use-yarn",
     "yarn": "4.1.0"
   },
@@ -24,14 +24,18 @@
     "build": "yarn workspaces foreach --all --exclude next-app run build",
     "upgrade:tamagui": "yarn up '*tamagui*'@latest '@tamagui/*'@latest react-native-web-lite@latest",
     "upgrade:tamagui:canary": "yarn up '*tamagui*'@canary '@tamagui/*'@canary react-native-web-lite@canary",
-    "check-deps": "check-dependency-version-consistency ."
+    "check-deps": "check-dependency-version-consistency .",
+    "test": "cd apps/next && npm test",
+    "test:auth": "cd apps/next && npm test tests/auth/role-assignment-simple.test.ts",
+    "test:e2e": "cd apps/next && npm run test:e2e",
+    "fix-auth": "bash scripts/test-and-fix.sh"
   },
   "resolutions": {
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
     "react-refresh": "^0.14.0",
     "react-native-svg": "15.3.0",
-    "react-native-web": "^0.19.12"
+    "react-native-web": "^0.20.0"
   },
   "dependencies": {
     "@babel/runtime": "^7.26.9",
@@ -41,7 +45,7 @@
     "node-gyp": "^11.1.0",
     "prettier": "^3.5.2",
     "turbo": "^2.4.2",
-    "typescript": "^5.7.3"
+    "typescript": "^5.8.3"
   },
   "packageManager": "yarn@4.6.0",
   "devDependencies": {
