@@ -1,12 +1,18 @@
 /// <reference types="vitest" />
-import { defineConfig } from 'vite'
+import { defineConfig } from 'vitest/config'
 import path from 'path'
+import { fileURLToPath } from 'url'
+
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
 
 export default defineConfig({
   test: {
     globals: true,
     environment: 'node',
     setupFiles: ['./tests/setup.ts'],
+    exclude: ['**/node_modules/**', '**/e2e/**', '**/*.e2e.*', '**/*.spec.*'],
+    include: ['**/*.test.*'],
   },
   resolve: {
     alias: {
