@@ -1,11 +1,22 @@
+'use client'
+
 import React from 'react'
 import { signOut } from 'next-auth/react'
 import { NavItem, Text } from '@my/ui'
 
-export const NavitemLogout: React.FC = () => {
+type NavitemLogoutProps = {
+  handleOpenChange?: () => void
+}
+
+export const NavitemLogout: React.FC<NavitemLogoutProps> = ({ handleOpenChange }) => {
+  const handleSignOut = () => {
+    handleOpenChange && handleOpenChange()
+    signOut({ callbackUrl: '/' })
+  }
+
   return (
     <>
-      <NavItem onPress={() => signOut()}>
+      <NavItem onPress={handleSignOut}>
         <Text fontWeight={'bold'} color={'$red10Light'}>
           Sign Out
         </Text>
