@@ -4,7 +4,7 @@ import { Spinner, YStack } from '@my/ui'
 export const Loading: React.FC = () => {
   return (
     <YStack fullscreen display={'flex'} alignItems={'center'} justifyContent={'center'}>
-      <Spinner size="large" color="$orange10" />
+      <Spinner size="$2" color="$orange10" />
     </YStack>
   )
 }
